move get_intervals to macrostrat
